+++ bin/scripts/fatalError.php
@@ -2,7 +2,6 @@
 
 use Awesomite\ErrorDumper\ErrorDumper;
 
-$dumper = new ErrorDumper();
-$dumper->createDevHandler()->register();
+ErrorDumper::createDevHandler()->register();
 
 require_once __DIR__ . DIRECTORY_SEPARATOR . '_fatalError.php.fatal_error';