+++ bin/scripts/humanFriendlyErrorCode.php
@@ -2,11 +2,7 @@
 
 use Awesomite\ErrorDumper\ErrorDumper;
 
-$errorDumper = new ErrorDumper();
-$errorHandler = $errorDumper->createDevHandler();
-$errorHandler->register();
-
-$sandbox = $errorHandler->getErrorSandbox();
+ErrorDumper::createDevHandler()->register();
 
 trigger_error('Test error', E_USER_WARNING);
 return 'OK';